+++ web/app/themes/rebe/index.php
@@ -1,6 +1,2 @@
 <?php
-
-if (defined('ABSPATH')) {
-    update_option('template', get_option('template') . '/templates');
-}
-die("Kind Regards,\nRoots");
+include __DIR__ . '/dist/templates/index.php';