+++ web/app/themes/rebe/src/assets/scripts/projectConstants.js
@@ -1,4 +1,4 @@
 angular.module('projectConstants', [])
   .constant('apiUrl', '/api/')
-  .constant('templatesPath', '/app/themes/rebe/templates/'
+  .constant('templatesPath', '/app/themes/rebe/dist/templates/'
 );