+++ web/app/themes/rebe/src/lib/Sage/Asset.php
@@ -9,8 +9,6 @@ use Roots\Sage\Assets\ManifestInterface;
  */
 class Asset
 {
-    public static $dist = '/dist';
-
     /** @var ManifestInterface Currently used manifest */
     protected $manifest;
 
@@ -33,6 +31,6 @@ class Asset
     public function getUri()
     {
         $file = ($this->manifest ? $this->manifest->get($this->asset) : $this->asset);
-        return get_template_directory_uri() . self::$dist . '/' . $this->dir . '/' . $file;
+        return get_template_directory_uri() . '/' . $this->dir . '/' . $file;
     }
 }