+++ web/app/themes/rebe/src/lib/Sage/Template/WrapperInterface.php
@@ -1,4 +1,6 @@
-<?php namespace Roots\Sage\Template;
+<?php
+
+namespace Roots\Sage\Template;
 
 /**
  * Interface WrapperInterface
@@ -13,16 +15,15 @@ interface WrapperInterface
      *
      * @return string Wrapper template (FQPN of, e.g., `base-page.php`, `base.php`)
      */
-    public function getWrapper();
+    public function wrap();
 
     /**
      * @return string Wrapped template (FQPN of, e.g., `page.php`, `single.php`, `singular.php`)
      */
-    public function getTemplate();
+    public function unwrap();
 
     /**
      * @return string Slug of the WrapperInterface; e.g., `base`
      */
-    public function getSlug();
-
+    public function slug();
 }